edited layout and welcome page

+++ Blogger/Controllers/CommentController.cs
@@ -39,7 +39,6 @@ namespace Blogger.Controllers
             return RedirectToAction("Login", "Account"); 
         }
         [HttpDelete]
-        [HttpPost]
         public IActionResult DeleteComment(int id)
         {
             var comment = _context.Comments.Find(id);
@@ -49,11 +48,20 @@ namespace Blogger.Controllers
                 return NotFound();
             }
 
+            var username = User.Identity.Name;
+            var user = _context.Users.FirstOrDefault(x => x.Username == username);
+
+            if (comment.UserId != user.Id)
+            {
+                return Unauthorized(); 
+            }
+
             _context.Comments.Remove(comment);
             _context.SaveChanges();
 
             return RedirectToAction("AllPosts", "Post");
         }
+
